+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   if the `Backup.beforeExport`/`Backup.beforeImport` events are stopped;
 * the code of `Driver::export()` and `Driver::import()` methods has been absorbed by `BackupExport::export()` and
   `BackupImport::import()` and therefore those methods no longer exist;
+* removed useless `BackupExport::$config` property;
 * completely improved the `BackupImportTest` tests.
 
 ## 2.11 branch